Manutenção na estrutura do código
