+++ src/routes/routes.js
@@ -1,0 +1,5 @@
+const express = require('express');
+const router = express.Router();
+const caracteristicasRouter = require('../controller/caracController')
+
+router.get('/caracteristicas', caracteristicasRouter)
