+++ styles/listaStyle.css
@@ -2,8 +2,10 @@
 
 body {
     font-family: 'Poppins', sans-serif;
-     background: linear-gradient(-45deg, #ffce51, #ff7253, #fd5754);
+    background: linear-gradient(-45deg, #ffce51, #ff7253, #fd5754);
 }
+
+
 
 .container {
     display: flex;
@@ -13,21 +15,30 @@ body {
     height: 100vh;
 }
 
+.navbar {
+    width: 30px;
+    height: 5px;
+    background-color: black;
+    margin: 2px 0;
+    border-radius: 10px;
+}
+
 .content {
     display: flex;
     flex-wrap: wrap;
     justify-content: center;
     align-items: center;
-    align-content: flex-start;
+    align-content: center;
     width: 100%;
     height: 100vh;
+    transition: margin-left 0.3s ease;
 }
 
 .card {
     background: #fff;
     border-radius: 20px;
     box-shadow: 0px 2px 23px #000000;
-    width: 23%;
+    width: 19%;
     padding: 5px;
     margin: 10px;
     box-sizing: border-box;
@@ -51,12 +62,19 @@ body {
     }
 }
 
+@media (max-width: 576px) {
+    .content {
+        align-items: flex-start;
+        /* Ajuste para evitar que os cards se sobreponham em telas pequenas */
+    }
+}
+
 .mediaCard img {
     height: 200px;
     width: 100%;
 }
 
-.title {
+a .title {
     color: #2b3d59;
 }
 
@@ -75,4 +93,91 @@ body {
 
 .bottomCard {
     padding: 20px;
+}
+
+.sidebar {
+    width: 250px;
+    height: 100vh;
+    background: #9e5b4a;
+    position: fixed;
+    top: 0;
+    left: -250px;
+    transition: left 0.3s ease;
+    justify-content: center;
+    display: flex;
+}
+
+.sidebar-toggle {
+    width: 30px;
+    height: 30px;
+    border: none;
+    outline: none;
+    position: absolute;
+    top: 20px;
+    right: -30px;
+    cursor: pointer;
+    background: none;
+}
+
+.bar1, .bar2, .bar3 {
+  width: 30px;
+  height: 5px;
+  background-color: #333;
+  margin: 4px 10px;
+  transition: 0.4s;
+  border-radius: 5px;
+}
+
+.change .bar1 {
+  transform: translate(0, 11px) rotate(-45deg);
+}
+
+.change .bar2 {opacity: 0;}
+
+.change .bar3 {
+  transform: translate(0, -11px) rotate(45deg);
+}
+
+.menu {
+    list-style: none;
+    padding: 0;
+    margin: 50px 0;
+}
+
+.menu li {
+    margin-bottom: 10px;
+}
+
+.menu li a {
+    display: block;
+    color: #fff;
+    text-decoration: none;
+    padding: 10px;
+    transition: background 0.3s ease;
+}
+
+.menu li a:hover {
+    background: rgba(255, 255, 255, 0.2);
+}
+
+.show-sidebar {
+    left: 0;
+}
+
+@media (max-width: 768px) {
+    .sidebar {
+        width: 100%;
+        height: auto;
+        position: static;
+        background: none;
+        transition: none;
+    }
+
+    .sidebar-toggle {
+        display: none;
+    }
+
+    .content {
+        margin-left: 0;
+    }
 }
